+++ consensus/accept.go
@@ -2,41 +2,20 @@ package consensus
 
 import (
 	"errors"
-	"sort"
 
 	"github.com/NebulousLabs/Sia/encoding"
 )
 
-// Exported Errors
 var (
-	BadBlockErr       = errors.New("block is known to be invalid.")
-	BlockKnownErr     = errors.New("block exists in block map.")
-	EarlyTimestampErr = errors.New("block timestamp is too early, block is illegal.")
-	FutureBlockErr    = errors.New("block timestamp too far in future")
-	OrphanErr         = errors.New("block has no known parent")
-	LargeBlockErr     = errors.New("block is too large to be accepted")
-	MinerPayoutErr    = errors.New("miner payout sum does not equal block subsidy")
-	MissedTargetErr   = errors.New("block does not meet target")
+	BadBlockErr        = errors.New("block is known to be invalid")
+	BlockKnownErr      = errors.New("block exists in block map")
+	EarlyTimestampErr  = errors.New("block timestamp is too early")
+	FutureTimestampErr = errors.New("block timestamp too far in future")
+	OrphanErr          = errors.New("block has no known parent")
+	LargeBlockErr      = errors.New("block is too large to be accepted")
+	MinerPayoutErr     = errors.New("miner payout sum does not equal block subsidy")
+	MissedTargetErr    = errors.New("block does not meet target")
 )
-
-// earliestChildTimestamp returns the earliest timestamp that a child node
-// can have while still being valid. See section 'Timestamp Rules' in
-// Consensus.md.
-func (bn *blockNode) earliestChildTimestamp() Timestamp {
-	// Get the previous `MedianTimestampWindow` timestamps.
-	var intTimestamps []int
-	referenceNode := bn
-	for i := 0; i < MedianTimestampWindow; i++ {
-		intTimestamps = append(intTimestamps, int(referenceNode.block.Timestamp))
-		if referenceNode.parent != nil {
-			referenceNode = referenceNode.parent
-		}
-	}
-	sort.Ints(intTimestamps)
-
-	// Return the median of the sorted timestamps.
-	return Timestamp(intTimestamps[MedianTimestampWindow/2])
-}
 
 // checkMinerPayouts verifies that the sum of all the miner payouts is equal to
 // the block subsidy (which is the coinbase + miner fees).
@@ -80,11 +59,16 @@ func (s *State) validHeader(b Block) (err error) {
 		return OrphanErr
 	}
 
-	// Check the id meets the target. This is one of the earliest checks to
+	// Check the ID meets the target. This is one of the earliest checks to
 	// enforce that blocks need to have committed to a large amount of work
 	// before being verified - a DoS protection.
 	if !b.CheckTarget(parent.target) {
 		return MissedTargetErr
+	}
+
+	// Check that the block is the correct size.
+	if len(encoding.Marshal(b)) > BlockSizeLimit {
+		return LargeBlockErr
 	}
 
 	// If timestamp is too far in the past, reject and put in bad blocks.
@@ -96,7 +80,7 @@ func (s *State) validHeader(b Block) (err error) {
 	// will need to be responsible for resubmitting the block once it is no
 	// longer in the future.
 	if b.Timestamp > CurrentTimestamp()+FutureThreshold {
-		return FutureBlockErr
+		return FutureTimestampErr
 	}
 
 	// Verify that the miner payouts sum to the total amount of fees allowed to
@@ -106,9 +90,23 @@ func (s *State) validHeader(b Block) (err error) {
 		return
 	}
 
-	// Check that the block is the correct size.
-	if len(encoding.Marshal(b)) > BlockSizeLimit {
-		return LargeBlockErr
+	return
+}
+
+// addBlockToTree takes a block and a parent node, and adds a child node to the
+// parent containing the block. No validation is done.
+func (s *State) addBlockToTree(b Block) (err error) {
+	parentNode := s.blockMap[b.ParentID]
+	newNode := parentNode.newChild(b)
+
+	// Add the node to the block map
+	s.blockMap[b.ID()] = newNode
+
+	if newNode.heavierThan(s.currentBlockNode()) {
+		err = s.forkBlockchain(newNode)
+		if err != nil {
+			return
+		}
 	}
 
 	return
